+++ src/crc32.hpp
@@ -22,7 +22,7 @@ class CRC32 {
 public:
     CRC32(const uint32_t chunk_size);
 
-    int64_t crc32(const std::string& filename) const;
+    ssize_t crc32(const std::string& filename) const;
     
 private:
     uint32_t _crc32(uint32_t crc, const uint8_t *ptr, const size_t len) const;
